fixing pep8
